+++ sismic/clock/__init__.py
@@ -1,0 +1,3 @@
+from .clock import BaseClock, SimulatedClock, WallClock
+
+__all__ = ['BaseClock', 'SimulatedClock', 'WallClock']